Add the feature to render .obj models
Use tinyobjloader-c to load .obj model data, then use flat
shading to render all triangles of the model. No depth test.

It is worth noting that the .obj file uses the right-handed
coordinate system by default. This is also the coordinate
system used by most modeling software (such as Maya, blender, etc.).

Foolrenderer originally planned to use the DirectX style
left-hand coordinate system. However, additional steps are
required to process the model into a left-handed
coordinate system, which is a bit complicated, and it is not
the goal of this project to spend too much effort on
processing model data. So the current program
conceptually uses the right-handed coordinate system.

In the future, the renderer may completely use the OpenGL
style coordinate system.