Fixed an issue where triangles may not be drawn
Because the method of vector3_normalize() to judge the zero
vector is not accurate. As a result, when the triangle is very small,
calculate_triangle_normal() may return a zero vector. So some
triangles cannot be drawn.

For the same reason, the method of judging whether the matrix is
invertible in matrix4x4_inverse() has been modified.