Use depth test to calculate triangle visibility
Add framebuffer structure for recording color buffer, depth
buffer and buffer size. The draw_triangle() function is changed
to require a framebuffer structure as the rendering destination.
The program uses the depth buffer to solve the pixel visibility
problem, ensuring that triangles farther from the camera will not
cover triangles that are closer.

The buffer size is now maintained by the framebuffer object,
and the role of set_viewport() is now more consistent with
similar functions in OpenGL.