Add programmable rendering pipeline
Now the function of foolrenderer is closer to the modern graphics
API. The rendering process is implemented by user-written
shaders (vertex shader and fragment shader), which greatly
improves the flexibility of the program: there is no need to modify
the draw_triangle() function for new rendering effects.

The data passed by the vertex shader to the fragment shader can
be automatically interpolated, it is no longer necessary to
manually interpolate each data.