Modified texture related code
Add texture format, used to specify the format and data type of
the pixels stored in the texture. This will help expand the use of
textures in the future. The function to set pixel data is now
handled by set_texture_pixels(). Added some functions to get
texture configuration information.