Now draw_triangle() accepts incomplete framebuffer
The incomplete framebuffer here means that the framebuffer is
not attached with color buffer or depth buffer. Now passing an
incomplete framebuffer to the draw_triangle() function will not
cause a program error.

It is possible to use an incomplete framebuffer. For example,
in the shadow mapping algorithm, in the first step usually avoid
updating the color buffer, in order to save drawing time.