Implement the shadow mapping of directional light
Add shadow_casting shader to create the shadow map. Adjusted
the basic shader to support the shadow mapping algorithm.

Because the basic shader needs to interpolate
light_space_position, the maximum number of vector3 variables
in shader_context is increased to 3.

The program now renders the suzanne model by default.