Implement normal mapping
Add normal mapping support in basic shader. In this
implementation shading is still computed in view space,
the view space normal is calculated in the fragment shader.

In many resources on normal mapping, they tend to put
shading calculations in tangent space for performance
reason. But this way reduces the flexibility of the program,
e.g. environment mapping needs to be sampled in world
space. And the transformation to tangent space needs to
be careful, the matrix used in some resources is wrong.

In order to temporarily reduce program complexity, the
main() function has removed the code to read the
command line parameters. Assets used for rendering are
hardcoded in the code.