Fix the issue of link error under Windows
When using the GCC compiler it needs to manually link m library to use
the functions of math.h. This issue was resolved in commit e3205a7.
But this change resulted in link errors under Windows, because
Visual Studio's C/C++ toolchain doesn't have a library named m.

Modify CMakeLists.txt, do not link this library when m library is not
found.